backend/requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-dotenv
sqlalchemy
